requirements.txt
streamlit>=1.30.0
python-pptx>=0.6.23
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
